Revert "[Admin panel][Conversations] Refactoring"
This reverts commit ccef01544e59ae2e20b4bd1507da6ce6c812c282.
