+++ app/views/admin/community_conversations/show.haml
@@ -4,21 +4,21 @@
     = "-"
     = t("admin.communities.conversations.conversations")
 
-= render :partial => "admin/left_hand_navigation", :locals => { :links => admin_links_for(service.community) }
+= render :partial => "admin/left_hand_navigation", :locals => { :links => admin_links_for(@current_community) }
 
 .left-navi-section
   %h2
-    - starter = service.conversation.starter
+    - starter = conversation.starter
     - starter_name = PersonViewUtils.person_display_name_for_type(starter, 'first_name_with_initial')
     - starter_link = starter ? link_to(starter_name, person_path(starter.username)) : ''
-    - recipient = service.conversation.recipient
+    - recipient = conversation.recipient
     - recipient_name = PersonViewUtils.person_display_name_for_type(recipient, 'first_name_with_initial')
     - recipient_link = recipient ? link_to(recipient_name, person_path(recipient.username)) : ''
     = t("admin.communities.conversations.participants", starter: starter_link, author: recipient_link).html_safe
-    - listing = service.conversation.listing
+    - listing = conversation.listing
     - if listing
       - listing_title = listing.title || t("admin.communities.conversations.not_available")
       %h2=link_to_unless listing.deleted, listing_title, listing_path(listing.id)
   #messages
-    = render :partial => "conversations/message", :collection => service.conversation_messages, as: :message_or_action
+    = render :partial => "conversations/message", :collection => messages, as: :message_or_action
